Created 3d Layout of the map and placing example objects
I have create 3d Layout of the map, I have scaled it correctly to the player size. I have placed examples of 3d Object and they are compatible with unity, textures and size can be changed in unity. I have extend map a bit to have more space for adding the 3d Objects