Made more changes to the level
Made the map smaller and the corridors smaller which makes the house feel better to move around, i also started putting in assets to fill out the rooms.

The floor plan will not be changed anymore only thing i need to do is add more assets but for some reason i cant access all the assets at this moment in time.