Added Room indicator for player, added crosshair
+ Added Room Indicator UI
The player will now be able to see what room they are currently in, this makes it somewhat easier to navigate the house easier.

+ Added Crosshair
Player can now see what they are looking at directly.

+ Added Light Probes & Reflection Probes
Assets that are not static (such as the player, and collectable items) can now be affected by static lighting (baked lighting). Reflection probes are added to implement mirrors into the bathrooms.