Fixed Player able to collect non-collectables
~ Fixed player able to collect items not in object list
A NullPointerException would be thrown if the player picked up an object that wasn't collectable. Only collectables will now able to display they can be collected through highlight and UI.

~ Changed layer mask for furniture
Changing furniture layer mask to "Interactive" to prevent the player not being able to jump when standing on any furniture.

- Removed layer mask for object collection
Now comparing if the raycast hits an object that has "FindMe" tag, and if the object is in objects list to be able to collect item.