Optimized Map with Occlusion Culling, Added loading screen
+ Applied Unity's Occlusion Culling to the level
This will help with performance so that any object not in the camera's frustum, will not be rendered. (For example, an object behind the player, will not be rendered)

+ Added Loading Screen
This is to let the player know that the game is loading the next scene, and is not hung or has crashed.

~ Redesigned Settings screen
Removed Graphics and Audio sections.

~ Fixed Object UI design
Items now have a preferred & minimum height, this is so that all text can be the same size, rather than fit the size of the text to the bounding box set by the game's resolution.

~ Fixed Settings screen not closing when pressing Esc
Settings screen would not close if the escape key was pressed to resume the game. 

- Removed Old map
Removed old map scene to save up on build size and old scene being deprecated.