Implemented Minimap to UI, Focused collect UI on object
+ Added Minimap into UI
The player is now presented with a small minimap in the top right of the screen to show the current location of the player.

~ Modified Collection UI to focus on Object position
Collection UI will now be focused on object position, to show to user what key to press to collect, so that they won't need to look around to screen too much.

~ Fixed smartphone not having a collider
Smartphone in the office didn't have a collider - causing a bug where the game couldn't be winnable if the phone was one of the objects.