Made Scripts modular and available as Prefabs
~ Essential components have been made into Prefabs
The new prefabs can be simply dragged into the Hierarchy (Ensure that you bind the UI gameobject onto the UI field in the GameManager script in the Inspector). Everything should work correctly, although, if any error occured (Issue running play mode, or Exceptions in the console), please make an issue, and label it as bug.