Added star system, changed time mechanic
+ Added star system
Game score is now managed by a star system

~ Changed Time Mechanic
To work with the star system, the Time mechanic has been changed to work by increasing from 0, the game will automatically calculate how many stars the user will get based on the time, and how many items they've collected out of the total. The items will now take away 5 seconds from the time, however this may change so that for speedruns, it's impossible to get a time of 0 seconds.